Fix EnsureDeoptimized: can correctly replace stack
Mark HInvokeStaticOrDirect.environment_ as being live such that
the environment can be retained and the corresponding vregs can
be filled to the CodeInfo. Previously, ART abandons this
environment making us to fail in transforming the quick frame
to a shadow frame correctly. Specifically, we lost the register
mapping and cannot crawl the correct value from physical
registers or stack.