Choose from_code according to deoptimization kind
The from_code parameter of Thread::PushDeoptimizationContext(),
when set to true, causes the dex tiggerring deoptimization to be
executed twice. In artemis, we choose from_code=false to avoid
this. Also, in artemis, when explicitly deoptimize a method, we
directly reset its method entry such it won't be the case: just
removed from the OSR map, however, entry is not changed meaning
stilled JITted.